Refactor Dropdown component to match the formatting of the TextInput component
Refactor Dropdown component so that the Label and ValidationErrorMessage properties are integrated within the component itself. Moved the Dropdown component to the Forms folder to be alongside the TextInput component. Dropdown component also uses the CSS grid format that TextInput utilizes now.
