--- src/Dropdown/Dropdown.tsx
@@ -1,75 +1,0 @@
-import React, {
-  ReactElement,
-  FunctionComponent,
-  useContext,
-  ChangeEventHandler,
-} from 'react';
-import styled, { ThemeContext } from 'styled-components';
-import { BaseTheme } from '../Theme';
-
-interface DropdownOptionProps {
-  /** The label of the dropdown option */
-  label: string;
-  /** The value of the dropdown option */
-  value: string;
-  /** Specifies whether the dropdown option is disabled */
-  disabled?: boolean;
-}
-
-export interface DropdownProps {
-  /** The id tied to this dropdown field */
-  id?: string;
-  /** Function to call on change event */
-  onChange: ChangeEventHandler;
-  /** The name of the dropdown */
-  name: string;
-  /** An array of objects with the properties specified through DropdownOptions */
-  options: Array<DropdownOptionProps>;
-  /** The currently selected dropdown value */
-  value?: string;
-  /** The dropdown value selected by default */
-  defaultValue?: string;
-  /** If true, the dropdown is required to submit the form */
-  required?: boolean;
-}
-
-const StyledDropdown = styled.select`
-  color: ${({ theme }): string => (theme.color.text.dark)};
-  width: 100%;
-`;
-
-const Dropdown: FunctionComponent<DropdownProps> = (props): ReactElement => {
-  const {
-    id,
-    onChange,
-    name,
-    options,
-    value,
-    defaultValue,
-    required,
-  } = props;
-  const theme: BaseTheme = useContext(ThemeContext);
-  return (
-    <StyledDropdown
-      id={id}
-      onChange={onChange}
-      theme={theme}
-      name={name}
-      value={value}
-      defaultValue={defaultValue}
-      required={required}
-    >
-      {options.map((option) => (
-        <option
-          value={option.value}
-          key={option.value}
-          disabled={option.disabled}
-        >
-          {option.label}
-        </option>
-      ))}
-    </StyledDropdown>
-  );
-};
-
-export default Dropdown;
