--- src/Dropdown/index.ts
@@ -1,1 +1,0 @@
-export { default as Dropdown } from './Dropdown';
