+++ src/Forms/__tests__/Dropdown.test.tsx
@@ -1,0 +1,501 @@
+import React from 'react';
+import {
+  render,
+  fireEvent,
+  BoundFunction,
+  AllByRole,
+  GetByText,
+  QueryByText,
+} from 'test-utils';
+import { spy } from 'sinon';
+import { strictEqual, deepStrictEqual } from 'assert';
+import { POSITION } from 'Forms/Label';
+import Dropdown from '../Dropdown';
+
+describe('Dropdown', function () {
+  let getByText: BoundFunction<GetByText>;
+  let getAllByRole: BoundFunction<AllByRole>;
+  let queryByText: BoundFunction<QueryByText>;
+  let changeSpy;
+  const options = [
+    {
+      value: 'all',
+      label: 'All',
+    },
+    {
+      value: 'fall',
+      label: 'Fall',
+    },
+    {
+      value: 'spring',
+      label: 'Spring',
+      disabled: false,
+    },
+    {
+      value: 'summer',
+      label: 'Summer',
+      disabled: true,
+    },
+  ];
+  context('when errorMessage prop is present', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+          errorMessage="Error: Please select a semester."
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('renders the error message', function () {
+      getByText('Please select a semester', { exact: false });
+    });
+  });
+  context('when the errorMessage prop is not present', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole, queryByText } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('does not render the error message', function () {
+      strictEqual(queryByText('Error: Please select a semester', { exact: false }), null);
+    });
+  });
+  context('when isRequired prop is present', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+          isRequired
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('renders the asterisk (*) denoting that the field is required', function () {
+      getByText('*');
+    });
+  });
+  context('when isRequired prop is not present', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole, queryByText } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('does not render the asterisk (*), which denotes that the field is required', function () {
+      strictEqual(queryByText('*', { exact: false }), null);
+    });
+  });
+  context('when isLabelVisible prop is true', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole, queryByText } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+          isLabelVisible
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('has a visible label', function () {
+      getByText('semesters');
+    });
+  });
+  context('when isLabelVisible prop is false', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole, queryByText } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+          isLabelVisible={false}
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('the label value is hidden in the UI', function () {
+      const style = window.getComputedStyle(getByText('semesters'));
+      strictEqual(style.display, 'none');
+    });
+  });
+  context('when disabled prop is true', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+        />
+      ));
+    });
+    it('sets the disabled attribute to the corresponding option element to true', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const summerOption = Array.from(dropdown.options).filter((option) => option.value === 'summer')[0];
+      strictEqual(summerOption.disabled, true);
+    });
+  });
+  context('when disabled prop is false', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+        />
+      ));
+    });
+    it('sets the disabled attribute to the corresponding option element to false', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const springOption = Array.from(dropdown.options).filter((option) => option.value === 'spring')[0];
+      strictEqual(springOption.disabled, false);
+    });
+  });
+  context('when the disabled attribute to the corresponding option element is not provided', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+        />
+      ));
+    });
+    it('does not set the disabled attribute to the corresponding option element', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const fallOption = Array.from(dropdown.options).filter((option) => option.value === 'fall')[0];
+      strictEqual(fallOption.disabled, false);
+    });
+  });
+  context('when labelPosition prop is POSITION.TOP', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+          labelPosition={POSITION.TOP}
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('positions the label above the dropdown', function () {
+      const style = window.getComputedStyle(getByText('semesters').parentNode as HTMLElement);
+      strictEqual(style['grid-template-areas'], '"l l l" "i i i" "e e e"');
+    });
+  });
+  context('when labelPosition prop is POSITION.LEFT', function () {
+    beforeEach(function () {
+      changeSpy = spy();
+      ({ getByText, getAllByRole } = render(
+        <Dropdown
+          id="semesters"
+          options={options}
+          value="fall"
+          label="semesters"
+          name="semesters"
+          onChange={changeSpy}
+          labelPosition={POSITION.LEFT}
+        />
+      ));
+    });
+    it('renders', function () {
+      getByText('Spring');
+    });
+    it('calls the change handler when changed', function () {
+      fireEvent.change(document.getElementsByName('semesters')[0]);
+      strictEqual(changeSpy.callCount, 1);
+    });
+    it('contains the expected elements', function () {
+      const dropdownOptionsCount = getAllByRole('option').length;
+      strictEqual(dropdownOptionsCount, options.length);
+    });
+    it('renders the correct default value', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const defaultValue = dropdown.value;
+      strictEqual(defaultValue, 'fall');
+    });
+    it('contains the correct value, label, and disabled value for each dropdown item', function () {
+      const dropdown = document.getElementsByName('semesters')[0] as HTMLSelectElement;
+      const optionsFound = Array.from(dropdown.options).map((element) => ({
+        disabled: element.disabled,
+        value: element.value,
+        label: element.label,
+      }));
+      const optionsWithDefaults = options.map((option) => (
+        {
+          ...option,
+          disabled: Boolean(option.disabled),
+        }
+      ));
+      deepStrictEqual(optionsFound, optionsWithDefaults);
+    });
+    it('positions the label to the left of the dropdown', function () {
+      const style = window.getComputedStyle(getByText('semesters').parentNode as HTMLElement);
+      strictEqual(style['grid-template-areas'], '"l i i" ". e e"');
+    });
+  });
+});
