+++ src/Forms/index.ts
@@ -1,3 +1,4 @@
 export { default as ValidationErrorMessage } from './ValidationErrorMessage';
 export { default as TextInput } from './TextInput';
 export { default as NoteText } from './NoteText';
+export { default as Dropdown } from './Dropdown';
