+++ src/index.ts
@@ -4,7 +4,6 @@ export * from './Tabs';
 export * from './Theme';
 export * from './Tables';
 export * from './Layout';
-export * from './Dropdown';
 export * from './Icons';
 export * from './Forms';
 export * from './Headings';
